ci: use Xcode 15.2
